[APG-908] Add Healthy Sex Programme (HSP) referral endpoints and tests
